sys/sys_io.h: Memory mapped I/O APIs must be pointer-sized, not 32 bit!
The mm_reg_t type used to represent a MMIO address (e.g. in the
sys_write*() APIs) was defined to a uint32_t, which is obviously wrong
on 64 bit systems where devices can be mapped anywhere.

+++ include/sys/sys_io.h
@@ -17,7 +17,7 @@ extern "C" {
 #endif
 
 typedef uint32_t io_port_t;
-typedef uint32_t mm_reg_t;
+typedef uintptr_t mm_reg_t;
 typedef uintptr_t mem_addr_t;
 
 /* Port I/O functions */
